add update account hook
